Stash smarter
I hit an issue before this where switching would fail because I was in
the middle of a `git am`, so we would stash, which would no-op, and then
we would pop, which pulled a random stash off. This now actually checks
if stashing happened based on the previous count.