Change update command to not parse output
Sometimes updatepr can contain the output of the diff from the upstream
branch, which could be a lot after a rebase, and could contain some
crazy chars. This caused our output parsing to fail sometimes, but we
don't actually care about the output parsing at all for this command, so
we can just check the exit code instead.