Handle PR creation failure cleanup
This is a bit risky since it deletes the remote branch as well in this
case, but if you just passed an invalid argument, or don't have network
access, without this cleanup you would have to do that yourself before
submitting the PR again. It should always be safe because if the branch
already existed upstream it would likely fail to be pushed to.