Abort rebase if continue fails
If mergetool succeeds but somehow the conflicting files are still marked
as unresolved, it leaves the checkout in a bad state. Now it at least
aborts the rebase. I don't know how it's possible to have a successful
mergetool with unresolved conflicts but I haven't investigated. The 1
log I have seen shows the conflicts were resolved with rerere, but then
the continue still failed.